En masse test

+++ src/server/model/ServerModelShowingManager.java
@@ -44,6 +44,9 @@ public class ServerModelShowingManager implements ServerModelShowing
 
   @Override public void removeShowing(IShowing showing) throws ServerException
   {
+    if (!showingDAO.getAllShowings(showing.getMovie()).contains(showing)){
+      throw new ServerException("That Showing does not exist");
+    }
     showingDAO.removeShowing(showing);
   }
